requirements.txt
attr==0.3.2
attrs==22.1.0
ConfigParser==5.3.0
cryptography==38.0.4
Cython==0.29.32
deta==1.1.0
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
ipython==8.7.0
ipywidgets==8.0.4
Jinja2==3.1.2
jnius==1.1.0
keyring==23.13.1
lockfile==0.12.2
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.3.0
protobuf==4.21.12
pyodbc==4.0.35
pyOpenSSL==22.1.0
python-dotenv==0.21.0
pywin32==305
railroad==0.5.0
smartsheet_python_sdk==3.0.0
Sphinx==5.3.0
SQLAlchemy==1.4.45
st_clickable_images==0.0.3
streamlit==1.16.0
streamlit_autorefresh==0.0.1
streamlit_extras==0.2.4
streamlit_multipage==0.0.18
streamlit_permalink==0.3.0
streamlit_ws_localstorage==1.0.0
toml==0.10.2
tornado==6.2
trove_classifiers==2022.12.22
truststore==0.5.0
websockets==10.4
xmlrpclib==1.0.1
~treamlit==1.15.2
